server/requirements.txt
python-dotenv
Flask
openai
